Use icons instead of circles where able. Refactor maps a bit.
