+++ frontend/src/dispatch/Dispatch.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { Link } from 'raviger';
-import {  Col, ListGroup, Row } from 'react-bootstrap'
-import { CircleMarker, Map, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
+import { Col, ListGroup, Row } from 'react-bootstrap'
+import { CircleMarker, Tooltip as MapTooltip } from "react-leaflet";
 import L from "leaflet";
 import Moment from 'react-moment';
 import randomColor from "randomcolor";
-import { Legend } from "../components/Map";
+import Map, { countMatches, prettyText } from "../components/Map";
 import badge from "../static/images/badge-sheriff.png";
 import bandaid from "../static/images/band-aid-solid.png";
 import car from "../static/images/car-solid.png";
@@ -16,62 +16,6 @@ function Dispatch() {
 
   const [data, setData] = useState({dispatch_assignments: [], isFetching: false, bounds:L.latLngBounds([[0,0]])});
   const [mapState, setMapState] = useState({});
-
-  // Takes in animal size, species, and count and returns a pretty string combination.
-  const prettyText = (size, species, count) => {
-    if (count <= 0) {
-      return "";
-    }
-    var plural = ""
-    if (count > 1) {
-      plural = "s"
-    }
-
-    var size_and_species = size + " " + species + plural;
-    // Exception for horses since they don't need an extra species output.
-    if (species === 'horse') {
-      // Exception for pluralizing ponies.
-      if (size === 'pony' && count > 1) {
-        size_and_species = 'ponies'
-      }
-      else {
-        size_and_species = size + plural;
-      }
-    }
-
-    var text = count + " " + size_and_species;
-    return text;
-  }
-
-  // Counts the number of size/species matches for a service request by status.
-  const countMatches = (service_request) => {
-    var matches = {};
-
-    service_request.animals.forEach((animal) => {
-      if (!matches[[animal.species,animal.size]]) {
-        matches[[animal.species,animal.size]] = 1;
-      }
-      else {
-        matches[[animal.species,animal.size]] += 1;
-      }
-    });
-    return matches
-  }
-
-  // Show or hide list of DAs based on current map zoom
-  const onMove = event => {
-    for (const dispatch_assignment of data.dispatch_assignments) {
-      let hidden = true;
-      for (const service_request of dispatch_assignment.service_request_objects) {
-        if (mapState[dispatch_assignment.id].service_requests[service_request.id]) {
-          if (event.target.getBounds().contains(L.latLng(service_request.latitude, service_request.longitude))) {
-            hidden = false;
-          }
-        }
-      }
-      setMapState(prevState => ({ ...prevState, [dispatch_assignment.id]: {...prevState[dispatch_assignment.id], hidden:hidden} }));
-    }
-  }
 
   // Hook for initializing data.
   useEffect(() => {
@@ -94,7 +38,7 @@ function Dispatch() {
         response.data.forEach((dispatch_assignment, index) => {
           let sr_dict = {}
           for (const service_request of dispatch_assignment.service_request_objects) {
-            const matches = countMatches(service_request);
+            const matches = countMatches(service_request)[0];
             sr_dict[service_request.id] = {id:service_request.id, matches:matches, latitude:service_request.latitude, longitude:service_request.longitude, assigned_evac:service_request.assigned_evac.id, full_address:service_request.full_address};
             bounds.push([service_request.latitude, service_request.longitude]);
           }
@@ -132,12 +76,7 @@ function Dispatch() {
     </ListGroup>
     <Row className="d-flex flex-wrap">
       <Col xs={10} className="border rounded pl-0 pr-0 m-auto">
-        <Map className="d-block" bounds={data.bounds} onMoveEnd={onMove}>
-          <Legend position="bottomleft" metric={false} />
-          <TileLayer
-            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-            attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-          />
+        <Map bounds={data.bounds}>
           {data.dispatch_assignments.map(dispatch_assignment => (
           <span key={dispatch_assignment.id}>
             {dispatch_assignment.service_request_objects.map(service_request => (
