+++ frontend/src/dispatch/DispatchMap.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from 'react';
-import ReactDOMServer from 'react-dom/server';
 import axios from "axios";
 import { Link, navigate } from 'raviger';
 import { Form, Formik } from 'formik';
@@ -8,10 +7,9 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faBandAid, faBullseye, faCar, faCircle, faClipboardList, faExclamationCircle, faQuestionCircle, faTrailer
 } from '@fortawesome/free-solid-svg-icons';
-import { faCheckCircle, faQuestionCircle as faQuestionCircleDuo } from '@fortawesome/pro-duotone-svg-icons';
+
 import { faBadgeSheriff, faHomeAlt } from '@fortawesome/pro-solid-svg-icons';
-import { faHomeAlt as faHomeAltReg } from '@fortawesome/pro-regular-svg-icons';
-import { Circle, Map, Marker, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
+import { Circle, Marker, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
 import L from "leaflet";
 import badge from "../static/images/badge-sheriff.png";
 import bandaid from "../static/images/band-aid-solid.png";
@@ -19,7 +17,7 @@ import car from "../static/images/car-solid.png";
 import trailer from "../static/images/trailer-solid.png";
 import { Typeahead } from 'react-bootstrap-typeahead';
 import Moment from 'react-moment';
-import { Legend } from "../components/Map";
+import Map, { countMatches, Legend, prettyText, reportedMarkerIcon, SIPMarkerIcon, UTLMarkerIcon, checkMarkerIcon } from "../components/Map";
 import { Checkbox } from "../components/Form";
 import 'react-bootstrap-typeahead/css/Typeahead.css';
 import 'leaflet/dist/leaflet.css';
@@ -103,59 +101,6 @@ function Deploy() {
     }
   }
 
-  // Takes in animal size, species, and count and returns a pretty string combination.
-  const prettyText = (size, species, count) => {
-    if (count <= 0) {
-      return "";
-    }
-    var plural = ""
-    if (count > 1) {
-      plural = "s"
-    }
-
-    var size_and_species = size + " " + species + plural;
-    // Exception for horses since they don't need an extra species output.
-    if (species === 'horse') {
-      // Exception for pluralizing ponies.
-      if (size === 'pony' && count > 1) {
-        size_and_species = 'ponies'
-      }
-      if (size === 'unknown') {
-        size_and_species = 'horse' + plural
-      }
-      else {
-        size_and_species = size + plural;
-      }
-    }
-
-    var text = count + " " + size_and_species;
-    return text;
-  }
-
-  // Counts the number of size/species matches for a service request by status.
-  const countMatches = (service_request) => {
-    var matches = {};
-    var status_matches = {'REPORTED':{}, 'SHELTERED IN PLACE':{}, 'UNABLE TO LOCATE':{}};
-
-    service_request.animals.forEach((animal) => {
-      if (['REPORTED', 'SHELTERED IN PLACE', 'UNABLE TO LOCATE'].indexOf(animal.status) > -1) {
-        if (!matches[[animal.species,animal.size]]) {
-          matches[[animal.species,animal.size]] = 1;
-        }
-        else {
-          matches[[animal.species,animal.size]] += 1;
-        }
-        if (!status_matches[animal.status][[animal.species,animal.size]]) {
-          status_matches[animal.status][[animal.species,animal.size]] = 1;
-        }
-        else {
-          status_matches[animal.status][[animal.species,animal.size]] += 1;
-        }
-      }
-    });
-    return [matches, status_matches]
-  }
-
   // Show or hide list of SRs based on current map zoom
   const onMove = event => {
     let tempMapState = mapState;
@@ -171,60 +116,6 @@ function Deploy() {
     }
     setMapState(prevState => ({ ...prevState, tempMapState}));
   }
-
-  const reportedIconHTML = ReactDOMServer.renderToString(<FontAwesomeIcon icon={faExclamationCircle} className="icon-border" color="#ff4c4c" size="lg" />);
-  const reportedMarkerIcon = new L.DivIcon({
-    html: reportedIconHTML,
-    iconSize: [0, 0],
-    iconAnchor: [6, 9],
-    className: "reported-icon",
-    popupAnchor: null,
-    shadowUrl: null,
-    shadowSize: null,
-    shadowAnchor: null
-  });
-
-  const SIPIconHTML = ReactDOMServer.renderToString(
-    <span className="fa-layers">
-      <FontAwesomeIcon icon={faCircle} className="icon-border" color="#f5ee0f" size="lg" transform={'grow-2'} />
-      <FontAwesomeIcon icon={faHomeAlt} style={{color:"white"}} transform={'shrink-3 left-1'} size="lg" inverse />
-      <FontAwesomeIcon icon={faHomeAltReg} style={{color:"#444"}} transform={'shrink-3 left-1'} size="lg" inverse />
-    </span>
-  );
-  const SIPMarkerIcon = new L.DivIcon({
-    html: SIPIconHTML,
-    iconSize: [0, 0],
-    iconAnchor: [5, 9],
-    className: "SIP-icon",
-    popupAnchor: null,
-    shadowUrl: null,
-    shadowSize: null,
-    shadowAnchor: null
-  });
-
-  const UTLIconHTML = ReactDOMServer.renderToString(<FontAwesomeIcon icon={faQuestionCircleDuo} className="icon-border" size="lg" style={{"--fa-primary-color":'white', "--fa-secondary-color":'#5f5fff', "--fa-secondary-opacity": 1}}  />);
-  const UTLMarkerIcon = new L.DivIcon({
-    html: UTLIconHTML,
-    iconSize: [0, 0],
-    iconAnchor: [6, 9],
-    className: "UTL-icon",
-    popupAnchor: null,
-    shadowUrl: null,
-    shadowSize: null,
-    shadowAnchor: null
-  });
-
-  const checkIconHTML = ReactDOMServer.renderToString(<FontAwesomeIcon icon={faCheckCircle} className="icon-border" size="lg" style={{"--fa-primary-color":'white', "--fa-secondary-color":'green', "--fa-secondary-opacity": 1}} />);
-  const checkMarkerIcon = new L.DivIcon({
-    html: checkIconHTML,
-    iconSize: [0, 0],
-    iconAnchor: [6, 9],
-    className: "check-icon",
-    popupAnchor: null,
-    shadowUrl: null,
-    shadowSize: null,
-    shadowAnchor: null
-  });
 
   // Hook for initializing data.
   useEffect(() => {
@@ -380,12 +271,7 @@ function Deploy() {
             </div>
           </Col>
           <Col xs={10} className="border rounded pl-0 pr-0">
-            <Map className="d-block" style={{marginRight:"0px"}} bounds={data.bounds} onMoveEnd={onMove}>
-              <Legend position="bottomleft" metric={false} />
-              <TileLayer
-                url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-                attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-              />
+            <Map style={{marginRight:"0px"}} bounds={data.bounds} onMoveEnd={onMove}>
               {data.service_requests.map(service_request => (
                 <Marker
                   position={[service_request.latitude, service_request.longitude]}
