+++ frontend/src/dispatch/DispatchSummary.js
@@ -1,16 +1,17 @@
 import React, { useEffect, useState } from 'react';
+import ReactDOMServer from 'react-dom/server';
 import axios from "axios";
 import { Link } from 'raviger';
 import { Button, Card, Col, ListGroup, Modal, OverlayTrigger, Row, Tooltip } from 'react-bootstrap';
 import { Typeahead } from 'react-bootstrap-typeahead';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
-  faClipboardCheck, faClipboardList, faEdit, faMinusSquare, faPlusSquare
+  faCircle, faClipboardCheck, faClipboardList, faEdit, faMinusSquare, faPlusSquare, faExclamationCircle
 } from '@fortawesome/free-solid-svg-icons';
-import { CircleMarker, Map, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
+import { Marker, TileLayer, Tooltip as MapTooltip } from "react-leaflet";
 import L from "leaflet";
 import Moment from 'react-moment';
-import { Legend } from "../components/Map";
+import Map, { countMatches, prettyText, reportedMarkerIcon, SIPMarkerIcon, UTLMarkerIcon } from "../components/Map";
 import Header from '../components/Header';
 
 function DispatchSummary({id}) {
@@ -59,52 +60,6 @@ function DispatchSummary({id}) {
     });
   }
 
-  // Takes in animal size, species, and count and returns a pretty string combination.
-  const prettyText = (size, species, count) => {
-    if (count <= 0) {
-      return "";
-    }
-    var plural = ""
-    if (count > 1) {
-      plural = "s"
-    }
-
-    var size_and_species = size + " " + species + plural;
-    // Exception for horses since they don't need an extra species output.
-    if (species === 'horse') {
-      // Exception for pluralizing ponies.
-      if (size === 'pony' && count > 1) {
-        size_and_species = 'ponies'
-      }
-      if (size === 'unknown') {
-        size_and_species = 'horse' + plural
-      }
-      else {
-        size_and_species = size + plural;
-      }
-    }
-
-    var text = count + " " + size_and_species;
-    return text;
-  }
-
-  // Counts the number of size/species matches for a service request by status.
-  const countMatches = (service_request) => {
-    var matches = {};
-
-    service_request.animals.forEach((animal) => {
-      if (['REPORTED', 'SHELTERED IN PLACE', 'UNABLE TO LOCATE'].indexOf(animal.status) > -1) {
-        if (!matches[[animal.species,animal.size]]) {
-          matches[[animal.species,animal.size]] = 1;
-        }
-        else {
-          matches[[animal.species,animal.size]] += 1;
-        }
-      }
-    });
-    return matches
-  }
-
   // Hook for initializing data.
   useEffect(() => {
     let source = axios.CancelToken.source();
@@ -117,7 +72,7 @@ function DispatchSummary({id}) {
         const map_dict = {};
         const bounds = [];
         for (const service_request of response.data.service_request_objects) {
-          const matches = countMatches(service_request);
+          const matches = countMatches(service_request)[0];
           map_dict[service_request.id] = {matches:matches, has_reported_animals:service_request.reported_animals > 0, latitude:service_request.latitude, longitude:service_request.longitude};
           bounds.push([service_request.latitude, service_request.longitude]);
         }
@@ -224,22 +179,11 @@ function DispatchSummary({id}) {
       </Col>
       <Col className="border rounded pl-0 pr-0" style={{marginTop:"4px", marginRight:"15px", maxHeight:"311px"}}>
         <Map className="d-block dispatch-leaflet-container" bounds={data.bounds}>
-          <Legend position="bottomleft" metric={false} />
-          <TileLayer
-            url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-            attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-          />
           {data.service_request_objects.map(service_request => (
-            <CircleMarker
-              key={service_request.id}
-              center={{lat:service_request.latitude, lng: service_request.longitude}}
-              color="black"
-              weight="1"
-              // fillColor={mapState[service_request.id] ? mapState[service_request.id].color : ""}
-              fill={true}
-              fillOpacity="1"
-              onClick={() => window.open("/hotline/servicerequest/" + service_request.id, "_blank")}
-              radius={5}
+            <Marker
+              position={[service_request.latitude, service_request.longitude]}
+              icon={service_request.sheltered_in_place > 0 ? SIPMarkerIcon : service_request.unable_to_locate > 0 ? UTLMarkerIcon : reportedMarkerIcon}
+              // onClick={() => window.open("/hotline/servicerequest/" + service_request.id, "_blank")}
             >
               <MapTooltip autoPan={false}>
                 <span>
@@ -256,7 +200,7 @@ function DispatchSummary({id}) {
                   {service_request.full_address}
                 </span>
               </MapTooltip>
-            </CircleMarker>
+            </Marker>
           ))}
         </Map>
       </Col>
